Began refactoring some structure, implementing new CSV loading.
