+++ KioskApplication/src/main/java/controller/MainWindowController.java
@@ -14,7 +14,7 @@ import javafx.scene.control.Tab;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
 import utility.ApplicationScreen;
-import utility.Node.NodeFloor;
+import utility.node.NodeFloor;
 
 import java.io.IOException;
 import java.util.HashMap;
@@ -142,10 +142,10 @@ public class MainWindowController {
                     case "Map Builder":
                         switchToScreen(ApplicationScreen.MAP_BUILDER);
                         break;
-                    case "Request Manager":
+                    case "request Manager":
                         switchToScreen(ApplicationScreen.REQUEST_MANAGER);
                         break;
-                    case "Request Submit":
+                    case "request Submit":
                         switchToScreen(ApplicationScreen.REQUEST_SUBMITTER);
                         break;
                     case "Settings":
