+++ KioskApplication/src/main/java/controller/RequestSubmitterController.java
@@ -12,9 +12,9 @@ import javafx.scene.control.Tab;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import utility.ApplicationScreen;
-import utility.Request.Language;
-import utility.Node.NodeFloor;
-import utility.Request.RequestType;
+import utility.request.Language;
+import utility.node.NodeFloor;
+import utility.request.RequestType;
 
 import java.io.IOException;
 import java.util.LinkedList;
