+++ KioskApplication/src/main/java/database/objects/InterpreterRequest.java
@@ -1,7 +1,7 @@
 package database.objects;
 
-import utility.Request.Language;
-import utility.Request.RequestProgressStatus;
+import utility.request.Language;
+import utility.request.RequestProgressStatus;
 
 import java.sql.Timestamp;
