--- KioskApplication/src/main/java/database/template/CSVFormat.java
@@ -1,6 +1,0 @@
-package database.template;
-
-public class CSVFormat {
-    public static final String NODE_CSV_HEAD = "nodeID,xcoord,ycoord,floor,building,nodeType,longName,shortName,teamAssigned";
-    public static final String EDGE_CSV_HEAD = "edgeID,startNode,endNode";
-}
