+++ KioskApplication/src/main/java/utility/csv/CsvException.java
@@ -1,0 +1,7 @@
+package utility.csv;
+
+public class CsvException extends Exception {
+    public CsvException(String ex) {
+        super(ex);
+    }
+}
