+++ KioskApplication/src/main/java/utility/javaFXUtil.java
@@ -6,7 +6,7 @@ import javafx.util.Duration;
 public class javaFXUtil {
 
     /**
-     * Make a Node fade out, the input arguement is Duration...
+     * Make a node fade out, the input arguement is Duration...
      * e.g. someNode.makeFadeOut(Duration.millis(1000), someNode, 1, 0);
      * Above sets the transition's duration to be 1000 ms, and from opacity of 1 to 0 (== fade out)
      */
