+++ KioskApplication/src/main/java/utility/node/NodeFloor.java
@@ -1,4 +1,4 @@
-package utility.Node;
+package utility.node;
 
 public enum NodeFloor {
     LOWERLEVEL_2,
