+++ KioskApplication/src/main/java/utility/node/NodeType.java
@@ -1,4 +1,4 @@
-package utility.Node;
+package utility.node;
 
 public enum NodeType {
     HALL, ELEV, REST, STAI, DEPT, LABS, INFO, CONF, EXIT, RETL, SERV
