+++ KioskApplication/src/main/java/utility/request/FoodService.java
@@ -1,4 +1,4 @@
-package utility.Request;
+package utility.request;
 
 public enum FoodService {
     foodType1, foodType2, foodType3, foodType4, foodType5
