+++ KioskApplication/src/main/java/utility/request/Language.java
@@ -1,4 +1,4 @@
-package utility.Request;
+package utility.request;
 
 public enum Language {
     SPANISH, CHINESE, FRENCH, TAGALOG, VIETNAMESE, KOREAN, GERMAN, ARABIC, RUSSIAN, ITALIAN, PORTUGUESE, NONE
