+++ KioskApplication/src/main/java/utility/request/RequestProgressStatus.java
@@ -1,4 +1,4 @@
-package utility.Request;
+package utility.request;
 
 public enum RequestProgressStatus {
     TO_DO, IN_PROGRESS, DONE;
