+++ KioskApplication/src/main/java/utility/request/RequestType.java
@@ -1,4 +1,4 @@
-package utility.Request;
+package utility.request;
 
 public enum RequestType {
     INTERPRETER,
