+++ KioskApplication/src/main/resources/view/SettingsView.fxml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import com.jfoenix.controls.JFXButton?>
 <?import com.jfoenix.controls.JFXRadioButton?>
 <?import com.jfoenix.controls.JFXTabPane?>
 <?import javafx.geometry.Insets?>
@@ -25,7 +26,31 @@
             </Tab>
             <Tab fx:id="databaseTab" text="Database">
               <content>
-                <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0" />
+                <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
+                     <children>
+                        <VBox layoutX="428.0" layoutY="280.0" prefHeight="200.0" prefWidth="100.0" spacing="5.0" AnchorPane.bottomAnchor="15.0" AnchorPane.leftAnchor="15.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="15.0">
+                           <children>
+                              <Label prefHeight="30.0" prefWidth="217.0" text="Database">
+                                 <font>
+                                    <Font size="25.0" />
+                                 </font>
+                              </Label>
+                              <Label prefHeight="30.0" prefWidth="217.0" text="Saving &amp; Loading">
+                                 <font>
+                                    <Font size="25.0" />
+                                 </font>
+                                 <VBox.margin>
+                                    <Insets />
+                                 </VBox.margin>
+                                 <padding>
+                                    <Insets right="5.0" />
+                                 </padding>
+                              </Label>
+                              <JFXButton buttonType="RAISED" onAction="#readCSV" text="Read CSVs" />
+                              <JFXButton buttonType="RAISED" onAction="#saveCSV" text="Save Database" />
+                           </children>
+                        </VBox>
+                     </children></AnchorPane>
               </content>
             </Tab>
             <Tab fx:id="pathfindingTab" text="Pathfinding">
@@ -33,7 +58,7 @@
                   <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="1085.0">
                      <children>
                         <Label fx:id="errorMsg" alignment="CENTER" contentDisplay="CENTER" layoutX="270.0" layoutY="276.0" prefHeight="47.0" prefWidth="154.0" textAlignment="CENTER" textFill="#da0b0b" textOverrun="CLIP" wrapText="true" />
-                        <Label fx:id="SALabel" layoutX="14.0" layoutY="14.0" prefHeight="30.0" prefWidth="217.0" text="Search Algorithm">
+                        <Label fx:id="SALabel" layoutX="14.0" layoutY="14.0" prefHeight="30.0" prefWidth="217.0" text="Search Algorithm" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="15.0">
                            <font>
                               <Font size="25.0" />
                            </font>
